@@ -1,62 +1,46 @@
 {
   "name": "ch05_3",
-  "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@fontsource/material-icons": "^5.0.7",
-    "@reduxjs/toolkit": "^1.9.5",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.53",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
-    "chance": "^1.1.11",
-    "luxon": "^3.4.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^8.1.2",
-    "react-scripts": "5.0.1",
-    "redux": "^4.2.1",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.4.2",
-    "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@fontsource/material-icons": "^5.0.19",
+    "@reduxjs/toolkit": "^2.2.7",
+    "chance": "^1.1.12",
+    "luxon": "^3.5.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.1.2",
+    "redux": "^5.0.1",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^3.1.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.9.0",
     "@tailwindcss/line-clamp": "^0.4.4",
-    "@types/chance": "^1.1.4",
-    "@types/luxon": "^3.3.2",
-    "@types/redux-logger": "^3.0.9",
+    "@types/chance": "^1.1.6",
+    "@types/luxon": "^3.4.2",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/redux-logger": "^3.0.13",
     "@types/redux-thunk": "^2.1.0",
-    "autoprefixer": "^10.4.16",
-    "daisyui": "^3.7.7",
-    "postcss": "^8.4.30",
-    "tailwindcss": "^3.3.3"
+    "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.20",
+    "daisyui": "^4.12.10",
+    "eslint": "^9.9.0",
+    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+    "eslint-plugin-react-refresh": "^0.4.9",
+    "globals": "^15.9.0",
+    "postcss": "^8.4.43",
+    "tailwindcss": "^3.4.10",
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^8.0.1",
+    "vite": "^5.4.1"
   }
 }
