@@ -1,8 +1,8 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
   theme: {
     extend: {}
   },
-  plugins: [require('@tailwindcss/line-clamp'), require('daisyui')]
+  plugins: [require('daisyui')]
 }
