@@ -3,8 +3,7 @@ import logo from "./logo.svg";
 import "./App.css";
 
 function App() {
-  return 
-  <h1>Hello React!</h1>;
+  return <h1>Hello React!</h1>;
 }
 
 export default App;
