@@ -4,5 +4,5 @@ export default {
   theme: {
     extend: {}
   },
-  plugins: [require('@tailwindcss/line-clamp'), require('daisyui')]
+  plugins: [require('daisyui')]
 }
