@@ -6,9 +6,10 @@ const initializeStore = () => {
   // configureStore는 리듀서에서 반환한 새로운 상태를 store 객체로 관리
   const store = configureStore({
     reducer: rootReducer,
-    middleware: getDefaultMiddleware => {
-      return []
-    }
+    middleware: getDefaultMiddleware =>
+      getDefaultMiddleware({
+        serializableCheck: false
+      })
   })
   return store
 }
