@@ -30,7 +30,7 @@ export default function Login() {
       .catch(e => {})
   }, [])
   return (
-    <div className="flex flex-col min-h-screen bg-gray-100 border border-gray-300 shadow-xl rounded-xl">
+    <div className="flex flex-col min-h-screen border-gray-300 rounded-xl shadow-xl bg-gray-100 border">
       <div className="flex flex-col items-center justify-center flex-1 max-w-sm px-2 mx-auto">
         <div className="w-full px-6 py-8 text-black bg-white rounded shadow-md">
           <h1 className="mb-8 text-2xl text-center text-primary">Login</h1>
